feat: added type-safe AbilityOwner Handles with the necessary interface changes
- split the AbilityScript into its own file due to circular dependency (although perhaps it should have been put in its own file from the get go)
