+++ Source/CkAbility/Public/CkAbility/AbilityOwner/CkAbilityOwner_Utils.h
@@ -32,9 +32,9 @@ public:
     UFUNCTION(BlueprintCallable,
               Category = "Ck|Utils|Ability|Owner",
               DisplayName="[Ck][AbilityOwner] Add Feature")
-    static void
+    static FCk_Handle_AbilityOwner
     Add(
-        FCk_Handle InHandle,
+        FCk_Handle& InHandle,
         const FCk_Fragment_AbilityOwner_ParamsData& InParams);
 
     UFUNCTION(BlueprintPure,
@@ -42,14 +42,24 @@ public:
               DisplayName="[Ck][AbilityOwner] Has Feature")
     static bool
     Has(
-        FCk_Handle InHandle);
+        const FCk_Handle& InHandle);
 
     UFUNCTION(BlueprintPure,
-              Category = "Ck|Utils|Ability|Owner",
-              DisplayName="[Ck][AbilityOwner] Ensure Has Feature")
-    static bool
-    Ensure(
-        FCk_Handle InHandle);
+        Category = "Ck|Utils|AbilityOwner",
+        DisplayName="[Ck][AbilityOwner] Cast",
+        meta = (ExpandEnumAsExecs = "OutResult"))
+    static FCk_Handle_AbilityOwner
+    Cast(
+        const FCk_Handle&    InHandle,
+        ECk_SucceededFailed& OutResult);
+
+    UFUNCTION(BlueprintPure,
+        Category = "Ck|Utils|AbilityOwner",
+        DisplayName="[Ck][AbilityOwner] Handle -> AbilityOwner Handle",
+        meta = (CompactNodeTitle = "As AbilityOwnerHandle", BlueprintAutocast))
+    static FCk_Handle_AbilityOwner
+    Conv_HandleToAbilityOwner(
+        const FCk_Handle& InHandle);
 
 public:
     UFUNCTION(BlueprintPure,
@@ -57,7 +67,7 @@ public:
               DisplayName="[Ck][AbilityOwner] Has Ability")
     static bool
     Has_Ability(
-        FCk_Handle InAbilityOwnerEntity,
+        const FCk_Handle_AbilityOwner& InAbilityOwnerEntity,
         TSubclassOf<UCk_Ability_Script_PDA> InAbilityClass);
 
     UFUNCTION(BlueprintPure,
@@ -65,14 +75,14 @@ public:
               DisplayName="[Ck][AbilityOwner] Has Any Ability")
     static bool
     Has_Any(
-        FCk_Handle InAbilityOwnerEntity);
+        const FCk_Handle_AbilityOwner& InAbilityOwnerEntity);
 
     UFUNCTION(BlueprintPure,
               Category = "Ck|Utils|Ability|Owner",
               DisplayName="[Ck][AbilityOwner] Ensure Has Ability")
     static bool
     Ensure_Ability(
-        FCk_Handle InAbilityOwnerEntity,
+        const FCk_Handle_AbilityOwner& InAbilityOwnerEntity,
         TSubclassOf<UCk_Ability_Script_PDA> InAbilityClass);
 
     UFUNCTION(BlueprintPure,
@@ -80,23 +90,23 @@ public:
               DisplayName="[Ck][AbilityOwner] Ensure Has Any Ability")
     static bool
     Ensure_Any(
-        FCk_Handle InAbilityOwnerEntity);
+        const FCk_Handle_AbilityOwner& InAbilityOwnerEntity);
 
     UFUNCTION(BlueprintPure,
               Category = "Ck|Utils|Ability|Owner",
               DisplayName="[Ck][AbilityOwner] Get Ability",
               meta=(DevelopmentOnly))
-    static FCk_Handle
+    static FCk_Handle_Ability
     Get_Ability(
-        FCk_Handle InAbilityOwnerEntity,
+        const FCk_Handle_AbilityOwner& InAbilityOwnerEntity,
         TSubclassOf<UCk_Ability_Script_PDA> InAbilityClass);
 
     UFUNCTION(BlueprintPure,
               Category = "Ck|Utils|Ability|Owner",
               DisplayName="[Ck][AbilityOwner] Find Ability")
-    static FCk_Handle
+    static FCk_Handle_Ability
     Find_Ability(
-        FCk_Handle InAbilityOwnerEntity,
+        const FCk_Handle_AbilityOwner& InAbilityOwnerEntity,
         FGameplayTag InAbilityName);
 
     UFUNCTION(BlueprintPure,
@@ -104,60 +114,60 @@ public:
               DisplayName="[Ck][AbilityOwner] Get Ability Count")
     static int32
     Get_AbilityCount(
-        FCk_Handle InAbilityOwnerEntity);
+        const FCk_Handle_AbilityOwner& InAbilityOwnerEntity);
 
 public:
     UFUNCTION(BlueprintCallable,
               Category = "Ck|Utils|Ability|Owner",
               DisplayName="[Ck][AbilityOwner] For Each Ability",
               meta=(AutoCreateRefTerm="InOptionalPayload, InDelegate"))
-    static TArray<FCk_Handle>
+    static TArray<FCk_Handle_Ability>
     ForEach_Ability(
-        FCk_Handle InAbilityOwnerEntity,
+        const FCk_Handle_AbilityOwner& InAbilityOwnerEntity,
         ECk_AbilityOwner_ForEachAbility_Policy InForEachAbilityPolicy,
         const FInstancedStruct& InOptionalPayload,
         const FCk_Lambda_InHandle& InDelegate);
     static auto
     ForEach_Ability(
-        FCk_Handle InAbilityOwnerEntity,
+        const FCk_Handle_AbilityOwner& InAbilityOwnerEntity,
         ECk_AbilityOwner_ForEachAbility_Policy InForEachAbilityPolicy,
-        const TFunction<void(FCk_Handle)>& InFunc) -> void;
+        const TFunction<void(FCk_Handle_Ability&)>& InFunc) -> void;
 
     UFUNCTION(BlueprintCallable,
               Category = "Ck|Utils|Ability|Owner",
               DisplayName="[Ck][AbilityOwner] For Each Ability If",
               meta=(AutoCreateRefTerm="InOptionalPayload, InDelegate"))
-    static TArray<FCk_Handle>
+    static TArray<FCk_Handle_Ability>
     ForEach_Ability_If(
-        FCk_Handle InAbilityOwnerEntity,
+        const FCk_Handle_AbilityOwner& InAbilityOwnerEntity,
         ECk_AbilityOwner_ForEachAbility_Policy InForEachAbilityPolicy,
         const FInstancedStruct& InOptionalPayload,
         const FCk_Lambda_InHandle& InDelegate,
         const FCk_Predicate_InHandle_OutResult& InPredicate);
     static auto
     ForEach_Ability_If(
-        FCk_Handle InAbilityOwnerEntity,
+        const FCk_Handle_AbilityOwner& InAbilityOwnerEntity,
         ECk_AbilityOwner_ForEachAbility_Policy InForEachAbilityPolicy,
-        const TFunction<void(FCk_Handle)>& InFunc,
-        const TFunction<bool(FCk_Handle)>& InPredicate) -> void;
+        const TFunction<void(FCk_Handle_Ability)>& InFunc,
+        const TFunction<bool(FCk_Handle_Ability)>& InPredicate) -> void;
 
     UFUNCTION(BlueprintCallable,
               Category = "Ck|Utils|Ability|Owner",
               DisplayName="[Ck][AbilityOwner] For Each Ability with Status",
               meta=(AutoCreateRefTerm="InOptionalPayload, InDelegate"))
-    static TArray<FCk_Handle>
+    static TArray<FCk_Handle_Ability>
     ForEach_Ability_WithStatus(
-        FCk_Handle InAbilityOwnerEntity,
+        const FCk_Handle_AbilityOwner& InAbilityOwnerEntity,
         ECk_Ability_Status InStatus,
         ECk_AbilityOwner_ForEachAbility_Policy InForEachAbilityPolicy,
         const FInstancedStruct& InOptionalPayload,
         const FCk_Lambda_InHandle& InDelegate);
     static auto
     ForEach_Ability_WithStatus(
-        FCk_Handle InAbilityOwnerEntity,
+        const FCk_Handle_AbilityOwner& InAbilityOwnerEntity,
         ECk_Ability_Status InStatus,
         ECk_AbilityOwner_ForEachAbility_Policy InForEachAbilityPolicy,
-        const TFunction<void(FCk_Handle)>& InFunc) -> void;
+        const TFunction<void(FCk_Handle_Ability)>& InFunc) -> void;
 
 public:
     UFUNCTION(BlueprintPure,
@@ -165,29 +175,29 @@ public:
               DisplayName="[Ck][AbilityOwner] Get Active Tags")
     static FGameplayTagContainer
     Get_ActiveTags(
-        FCk_Handle InAbilityOwnerHandle);
+        const FCk_Handle_AbilityOwner& InAbilityOwnerHandle);
 
     UFUNCTION(BlueprintPure,
               Category = "Ck|Utils|Ability|Owner",
               DisplayName="[Ck][AbilityOwner] Get Active Tags With Count")
     static TMap<FGameplayTag, int32>
     Get_ActiveTagsWithCount(
-        FCk_Handle InAbilityOwnerHandle);
+        const FCk_Handle_AbilityOwner& InAbilityOwnerHandle);
 public:
     UFUNCTION(BlueprintCallable,
               Category = "Ck|Utils|Ability|Owner",
               DisplayName="[Ck][AbilityOwner] Request Give Ability")
-    static void
+    static FCk_Handle_AbilityOwner
     Request_GiveAbility(
-        FCk_Handle InAbilityOwnerHandle,
+        FCk_Handle_AbilityOwner& InAbilityOwnerHandle,
         const FCk_Request_AbilityOwner_GiveAbility& InRequest);
 
     UFUNCTION(BlueprintCallable,
               Category = "Ck|Utils|Ability|Owner",
               DisplayName="[Ck][AbilityOwner] Request Revoke Ability")
-    static void
+    static FCk_Handle_AbilityOwner
     Request_RevokeAbility(
-        FCk_Handle InAbilityOwnerHandle,
+        FCk_Handle_AbilityOwner& InAbilityOwnerHandle,
         const FCk_Request_AbilityOwner_RevokeAbility& InRequest);
 
     // NOTE: This is for development only. Use 'Request_SendEvent' to trigger Activation of Abilities
@@ -195,9 +205,9 @@ public:
               Category = "Ck|Utils|Ability|Owner",
               DisplayName="[Ck][AbilityOwner] Request Try Activate Ability",
               meta = (DevelopmentOnly))
-    static void
+    static FCk_Handle_AbilityOwner
     Request_TryActivateAbility(
-        FCk_Handle InAbilityOwnerHandle,
+        FCk_Handle_AbilityOwner& InAbilityOwnerHandle,
         const FCk_Request_AbilityOwner_ActivateAbility& InRequest);
 
     // NOTE: This is for development only. Use 'Request_SendEvent' to trigger Deactivation of Abilities
@@ -205,52 +215,52 @@ public:
               Category = "Ck|Utils|Ability|Owner",
               DisplayName="[Ck][AbilityOwner] Request Deactivate Ability",
               meta = (DevelopmentOnly))
-    static void
+    static FCk_Handle_AbilityOwner
     Request_DeactivateAbility(
-        FCk_Handle InAbilityOwnerHandle,
+        FCk_Handle_AbilityOwner& InAbilityOwnerHandle,
         const FCk_Request_AbilityOwner_DeactivateAbility& InRequest);
 
     UFUNCTION(BlueprintCallable,
               Category = "Ck|Utils|Ability|Owner",
               DisplayName = "[Ck][AbilityOwner] Request Send Ability Event")
-    static void
+    static FCk_Handle_AbilityOwner
     Request_SendAbilityEvent(
-        FCk_Handle InAbilityOwnerHandle,
+        FCk_Handle_AbilityOwner& InAbilityOwnerHandle,
         const FCk_Request_AbilityOwner_SendEvent& InRequest);
 
 public:
     UFUNCTION(BlueprintCallable,
               Category = "Ck|Utils|Ability|Owner",
               DisplayName = "[Ck][AbilityOwner] Bind To OnEvents")
-    static void
+    static FCk_Handle_AbilityOwner
     BindTo_OnEvents(
-        FCk_Handle InAbilityOwnerHandle,
+        FCk_Handle_AbilityOwner& InAbilityOwnerHandle,
         ECk_Signal_BindingPolicy InBehavior,
         const FCk_Delegate_AbilityOwner_Events& InDelegate);
 
     UFUNCTION(BlueprintCallable,
               Category = "Ck|Utils|Ability|Owner",
               DisplayName = "[Ck][AbilityOwner] Unbind From OnEvents")
-    static void
+    static FCk_Handle_AbilityOwner
     UnbindFrom_OnEvents(
-        FCk_Handle InAbilityOwnerHandle,
+        FCk_Handle_AbilityOwner& InAbilityOwnerHandle,
         const FCk_Delegate_AbilityOwner_Events& InDelegate);
 
     UFUNCTION(BlueprintCallable,
               Category = "Ck|Utils|Ability|Owner",
               DisplayName = "[Ck][AbilityOwner] Bind To OnTagsUpdated")
-    static void
+    static FCk_Handle_AbilityOwner
     BindTo_OnTagsUpdated(
-        FCk_Handle InAbilityOwnerHandle,
+        FCk_Handle_AbilityOwner& InAbilityOwnerHandle,
         ECk_Signal_BindingPolicy InBehavior,
         const FCk_Delegate_AbilityOwner_OnTagsUpdated& InDelegate);
 
     UFUNCTION(BlueprintCallable,
               Category = "Ck|Utils|Ability|Owner",
               DisplayName = "[Ck][AbilityOwner] Unbind From OnTagsUpdated")
-    static void
+    static FCk_Handle_AbilityOwner
     UnbindFrom_OnTagsUpdated(
-        FCk_Handle InAbilityOwnerHandle,
+        FCk_Handle_AbilityOwner& InAbilityOwnerHandle,
         const FCk_Delegate_AbilityOwner_OnTagsUpdated& InDelegate);
 
 private:
@@ -265,7 +275,7 @@ private:
               Category = "Ck|Utils|Ability|Owner")
     static FCk_Request_AbilityOwner_ActivateAbility
     Make_Request_ActivateAbility_ByEntity(
-        FCk_Handle                    InAbilityEntity,
+        const FCk_Handle_Ability& InAbilityEntity,
         FCk_Ability_ActivationPayload InActivationPayload);
 
     UFUNCTION(BlueprintPure,
@@ -278,7 +288,7 @@ private:
               Category = "Ck|Utils|Ability|Owner")
     static FCk_Request_AbilityOwner_DeactivateAbility
     Make_Request_DeactivateAbility_ByEntity(
-        FCk_Handle InAbilityEntity);
+        const FCk_Handle_Ability& InAbilityEntity);
 
     UFUNCTION(BlueprintPure,
               Category = "Ck|Utils|Ability|Owner")
@@ -290,7 +300,7 @@ private:
               Category = "Ck|Utils|Ability|Owner")
     static FCk_Request_AbilityOwner_RevokeAbility
     Make_Request_RevokeAbility_ByEntity(
-        FCk_Handle InAbilityEntity);
+        const FCk_Handle_Ability& InAbilityEntity);
 };
 
 // --------------------------------------------------------------------------------------------------------------------
